refactor: move base repository contracts to 'Base/' directory
